// RRC.UI/src/app/media/calendar-management/model/calendar-management.model.ts
export class CalendarManagement {
  ReferenceNumber: string;
  EventRequestor: string;
  EventType: number;
  EventDetails: string;
  DateFrom: any;
  DateTo: any;
  AllDayEvents: boolean;
  Location: number;
  City: number;
  ApproverID: number;
  CreatedBy: number;
  CreatedDateTime: any;
  Action: string;
  Comments: string;
  ParentReferenceNumber: string;
  Status: any;
  UpdatedBy?:any;
  UpdatedDateTime?:any;
  CalendarID?:any;
}